rename the file containing the default models
