+++ tests/scorers/test_similarity_scorer.py
@@ -3,7 +3,7 @@ import random
 import pytest
 
 import weave
-from weave.scorers.llm_utils import OPENAI_DEFAULT_EMBEDDING_MODEL
+from weave.scorers.default_models import OPENAI_DEFAULT_EMBEDDING_MODEL
 from weave.scorers.similarity_scorer import (
     EmbeddingSimilarityScorer,
 )
