+++ weave/scorers/guardrails/prompt_injection_guardrail.py
@@ -4,12 +4,12 @@ from litellm import acompletion
 from pydantic import BaseModel
 
 import weave
+from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
 from weave.scorers.guardrails.prompts import (
     PROMPT_INJECTION_GUARDRAIL_SYSTEM_PROMPT,
     PROMPT_INJECTION_SURVEY_PAPER_SUMMARY,
 )
 from weave.scorers.llm_scorer import LLMScorer
-from weave.scorers.llm_utils import OPENAI_DEFAULT_MODEL
 from weave.scorers.utils import stringify
