+++ weave/scorers/hallucination_scorer.py
@@ -2,8 +2,8 @@ from litellm import acompletion
 from pydantic import BaseModel, Field
 
 import weave
+from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
 from weave.scorers.llm_scorer import LLMScorer
-from weave.scorers.llm_utils import OPENAI_DEFAULT_MODEL
 from weave.scorers.utils import stringify
 
 DEFAULT_HALLUCINATION_SYSTEM_PROMPT = """
