+++ weave/scorers/ragas_scorer.py
@@ -6,8 +6,8 @@ from litellm import acompletion
 from pydantic import BaseModel, Field
 
 import weave
+from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
 from weave.scorers.llm_scorer import LLMScorer
-from weave.scorers.llm_utils import OPENAI_DEFAULT_MODEL
 
 
 class EntityExtractionResponse(BaseModel):
